midi/requirements.txt
PyQt5
pyserial
python-rtmidi
